requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
Pillow
langchain-community
langchain-core
langchain-ollama
langchain-google-genai
python-dotenv